Added VBA Script
This is the same VBA script I used for each worksheet in both the alphabetical_testing and multiple_year_stock_data files. I successfully ran it on most of the sheets in the alphabetical_testing file (included my solved version of this file), except for the sheet titled P. The script ran successfully on year 2016 in the multiple_year_stock_data file, but returned errors when run on 2015 & 2014.